+++ src/context/Context.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+const Context = createContext();
+export default Context;
